pyproject.toml
[tool.poetry]
name = "qollab-grade-service"
version = "0.1.0"
description = ""
authors = ["Bernar Berdikul <bernar.berdikul@erg.kz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
sqlmodel = "^0.0.14"
uvicorn = "^0.25.0"
alembic = "^1.13.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
httpx = "^0.26.0"
orjson = "^3.9.10"
asyncpg = "^0.29.0"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
